[loong64][mips64][liftoff] Sign extend the return value when Call C
In the simulator, if return value of the C function is int32, it may
not be properly sign-extended; so we need to sign extend it in
liftoff.

Besides, sign-extend the result of DIVU in MIPS.

Bug: chromium:1451290, chromium:1454668

+++ src/execution/mips64/simulator-mips64.cc
@@ -4164,8 +4164,8 @@ void Simulator::DecodeTypeRegisterSPECIAL() {
           if (rt_u() != 0) {
             uint32_t rt_u_32 = static_cast<uint32_t>(rt_u());
             uint32_t rs_u_32 = static_cast<uint32_t>(rs_u());
-            set_register(LO, rs_u_32 / rt_u_32);
-            set_register(HI, rs_u_32 % rt_u_32);
+            set_register(LO, static_cast<int32_t>(rs_u_32 / rt_u_32));
+            set_register(HI, static_cast<int32_t>(rs_u_32 % rt_u_32));
           }
         }
       }
